[Security Solution] move field renderers to explore network folder as they're not used in Timeline anymore
