+++ x-pack/solutions/security/plugins/security_solution/public/explore/network/components/field_renderers/field_renderers.test.tsx
@@ -9,31 +9,31 @@ import React from 'react';
 // Necessary until components being tested are migrated of styled-components https://github.com/elastic/kibana/issues/219037
 import 'jest-styled-components';
 import { render, screen } from '@testing-library/react';
-import { TestProviders } from '../../../common/mock';
-import { getEmptyValue } from '../../../common/components/empty_value';
+import { TestProviders } from '../../../../common/mock';
+import { getEmptyValue } from '../../../../common/components/empty_value';
 import {
   autonomousSystemRenderer,
   hostIdRenderer,
   hostNameRenderer,
   locationRenderer,
-  whoisRenderer,
   reputationRenderer,
+  whoisRenderer,
 } from './field_renderers';
-import { mockData } from '../../../explore/network/components/details/mock';
-import type { AutonomousSystem } from '../../../../common/search_strategy';
-import { FlowTarget } from '../../../../common/search_strategy';
+import { mockData } from '../details/mock';
+import type { AutonomousSystem } from '../../../../../common/search_strategy';
+import { FlowTarget } from '../../../../../common/search_strategy';
 import type { HostEcs } from '@kbn/securitysolution-ecs';
 import { mockGetUrlForApp } from '@kbn/security-solution-navigation/mocks/context';
-import { SourcererScopeName } from '../../../sourcerer/store/model';
+import { SourcererScopeName } from '../../../../sourcerer/store/model';
 
-jest.mock('../../../common/lib/kibana');
+jest.mock('../../../../common/lib/kibana');
 jest.mock('@kbn/security-solution-navigation/src/context');
 mockGetUrlForApp.mockImplementation(
   (appId: string, options?: { path?: string; deepLinkId?: boolean }) =>
     `${appId}/${options?.deepLinkId ?? ''}${options?.path ?? ''}`
 );
 
-jest.mock('../../../common/hooks/use_get_field_spec');
+jest.mock('../../../../common/hooks/use_get_field_spec');
 
 const mockHost: HostEcs = mockData.complete.host as HostEcs;
